Widen SidecarBuilder sortBufferBytes arg to long
SourcelessSpillConfig.sortBufferBytes() returns long. SegmentTermIndex
was narrowing to int via (int) cast before passing to the builder, which
silently truncated (and could go negative) for configured values above
Integer.MAX_VALUE, falling through to the default. Widen the ctor arg to
long; clamp to [MIN_BUFFER_SIZE_MB, Integer.MAX_VALUE] MiB inside the
builder (Lucene BufferSize itself is int-MiB-bounded).

+++ SearchSnapshotExtractor/src/main/java/org/opensearch/migrations/bulkload/lucene/SegmentTermIndex.java
@@ -138,7 +138,7 @@ public class SegmentTermIndex implements AutoCloseable {
         // streamFieldPostings or buildAndOpenReader throws. On the happy path
         // buildAndOpenReader flips the builder's closed flag, so the subsequent
         // close() is a no-op and does not touch the files SidecarReader now owns.
-        try (SidecarBuilder builder = new SidecarBuilder(fieldSpillDir, (int) sortBufferBytes, maxDoc)) {
+        try (SidecarBuilder builder = new SidecarBuilder(fieldSpillDir, sortBufferBytes, maxDoc)) {
             reader.streamFieldPostings(fieldName, builder);
             return builder.buildAndOpenReader();
         }
